requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pathlib>=1.0.1
streamlit>=1.28.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0